requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
Django==5.2.1
django-cors-headers==4.7.0
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
priority==1.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
redis==6.0.0
service-identity==24.2.0
setuptools==80.4.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
zope.interface==7.2
